requirements.txt
psycopg2-binary
pydantic
rich
